updated run.sh and print_pattern.c

+++ C/hacker_rank/run.sh
@@ -1,14 +1,23 @@
 #!/usr/bin/env bash
 
 if [ $# -eq 0 ]; then
-    echo "Enter the filename to compile"
+    echo "+----------------------------+"
+    echo "| Usage: ./run.sh <filename> |"
+    echo "+----------------------------+"
+    echo "| Example:                   |"
+    echo "|          ./run.sh main.c   |"
+    echo "+----------------------------+"
     exit
 fi
 
 filename=$1
 exe_name=${filename%.*}
-gcc -o $exe_name $filename -Wall -Wextra -lm
+if [[ -f "$exe_name" && "$exe_name" -nt "$filename" ]];then
+    echo "Skipping compilation, file already compiled"
+else
+    gcc -o $exe_name $filename -Wall -Wextra -lm
+    ./$exe_name
 
-./$exe_name
-
-rm $exe_name
+    rm $exe_name
+fi
+echo "Done!"
